web/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib==1.7.4
requests==2.32.4
faker==37.8.0
